requirements.txt
orchestral-ai>=1.3.0
python-dotenv>=1.1.1
